+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithMultipleCtorsRepository.java
@@ -34,10 +34,10 @@ public final class Arez_RepositoryWithMultipleCtorsRepository extends Repository
 
   @Nonnull
   @Override
-  public RepositoryWithMultipleCtors create(@Nonnull final String packageName, @Nonnull final String name) {
+  public RepositoryWithMultipleCtors create() {
     Guards.invariant( () -> !$$arez$$_disposed, () -> "Method invoked on invalid component 'RepositoryWithMultipleCtorsRepository'" );
     try {
-      return this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? "RepositoryWithMultipleCtorsRepository.create_packageName_name" : null, true, () -> super.create(packageName,name), packageName, name );
+      return this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? "RepositoryWithMultipleCtorsRepository.create" : null, true, () -> super.create() );
     } catch( final RuntimeException $$arez$$_e ) {
       throw $$arez$$_e;
     } catch( final Exception $$arez$$_e ) {
@@ -51,10 +51,10 @@ public final class Arez_RepositoryWithMultipleCtorsRepository extends Repository
 
   @Nonnull
   @Override
-  public RepositoryWithMultipleCtors create() {
+  public RepositoryWithMultipleCtors create(@Nonnull final String packageName, @Nonnull final String name) {
     Guards.invariant( () -> !$$arez$$_disposed, () -> "Method invoked on invalid component 'RepositoryWithMultipleCtorsRepository'" );
     try {
-      return this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? "RepositoryWithMultipleCtorsRepository.create_" : null, true, () -> super.create() );
+      return this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? "RepositoryWithMultipleCtorsRepository.create_packageName_name" : null, true, () -> super.create(packageName,name), packageName, name );
     } catch( final RuntimeException $$arez$$_e ) {
       throw $$arez$$_e;
     } catch( final Exception $$arez$$_e ) {